# Fix close surfaces bug