@eepeterson 2023-07-22
Hi @egor1abs thanks for looking into this and proposing a fix! Does the `merge_surfaces` property of the `Geometry` class serve the purpose you're looking for? It removes redundant surfaces from the geometry before exporting to xml.