@egor1abs 2023-07-22
> Hi @egor1abs thanks for looking into this and proposing a fix! Does the `merge_surfaces` property of the `Geometry` class serve the purpose you're looking for? It removes redundant surfaces from the geometry before exporting to xml.

Hi, thanks for your answer!
I'm using xml files transferred by the converter from MCNP to OpenMC made by @paulromano. Maybe that is indeed the problem, will definitely check it out!