@paulromano 2023-07-24
Thanks for reporting this @egor1abs! In this case, I can add a new option in the MCNP converter to merge surfaces (right now it does not do it by default).