@paulromano 2023-07-24
@egor1abs I just updated the `openmc_mcnp_adapter` to now remove redundant surfaces by default. Give it a shot and if you still have issues, let me know!